requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pydantic>=2.0.0
pyyaml>=6.0
click>=8.0.0
faker>=18.0.0
prometheus-client>=0.16.0
jsonschema>=4.0.0
psutil>=5.9.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
pre-commit>=3.0.0
